GH#30241: remove marketing performance Qlty debt
