+++ .agents/scripts/_performance_contract_values.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3
+# SPDX-License-Identifier: MIT
+# SPDX-FileCopyrightText: 2025-2026 Marcus Quinn
+"""Compatibility exports for performance contract value helpers."""
+
+from _performance_contract_scalar import *  # noqa: F403
+from _performance_contract_wire import *  # noqa: F403
